updated player
AAAAAAAAAAA

+++ Assets/Scripts/LevelManager.cs
@@ -10,8 +10,8 @@ public class LevelManager : MonoBehaviour
     private float timer;
     public static int level;
     public GridSetup grid;
-    public static int maxScore;
-    public static int currentScore;
+    public static float maxScore;
+    public static float currentScore;
     // Start is called before the first frame update
     void Start()
     {
